+++ tests/realtikvtest/importintotest3/BUILD.bazel
@@ -15,6 +15,8 @@ go_test(
     race = "on",
     shard_count = 2,
     deps = [
+        "//pkg/config",
+        "//pkg/config/deploymode",
         "//pkg/config/kerneltype",
         "//pkg/dxf/framework/proto",
         "//pkg/dxf/framework/storage",
